+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2-beta.19] - 2022-11-11
+
+### Added
+
+-   Added `vv-nav` component for navigation and tabs.
+
+### Changed
+
+-   docs for `vv-badge` and `vv-avatar` components
+
 ## [0.1.2-beta.18] - 2022-11-10
 
 ### Changed
@@ -114,6 +124,7 @@ All notable changes to this project will be documented in this file.
 -   vv-card
 -   vv-text
 
+[0.1.2-beta.19]: https://github.com/volverjs/style/compare/v0.1.2-beta.18...v0.1.2-beta.19
 [0.1.2-beta.18]: https://github.com/volverjs/style/compare/v0.1.2-beta.17...v0.1.2-beta.18
 [0.1.2-beta.17]: https://github.com/volverjs/style/compare/v0.1.2-beta.16...v0.1.2-beta.17
 [0.1.2-beta.16]: https://github.com/volverjs/style/compare/v0.1.2-beta.15...v0.1.2-beta.16